Fix Get entity method issue and add EquipmetRepository unit test

+++ Challenge/Challenge.Dal/Interfaces/IRepository.cs
@@ -8,7 +8,7 @@ namespace Challenge.Dal.Interfaces
 {
     public interface IRepository<TEntity> where TEntity : class
     {
-        TEntity Get(int id);
+        TEntity Get(decimal id);
         IEnumerable<TEntity> GetAll();
         void Add(TEntity entity);
         void Remove(TEntity entity);
